Time Nodes and Connections
Save time to the progress file

Display Time, Node count and Connection count in the level menu if the level is solved.

This can later be used for level leaderboards.

To be added: Confirm overwritting
Right now, with this, when a level is submited it automatically overwrites the previous save, but the user may not want this, a new pop up is necessary.